+++ changelog.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## NEXT
+
+- Added the `cai_causal_graph.time_series_graph.TimeSeriesGraph.get_nodes_at_lag` and
+  `cai_causal_graph.time_series_graph.TimeSeriesGraph.get_contemporaneous_nodes` methods to the
+  `cai_causal_graph.time_series_graph.TimeSeriesGraph` class to get the nodes at a given lag and the contemporaneous
+  nodes of the provided node, respectively.
+- General improvements to several `from_*` methods in the `cai_causal_graph.time_series_graph.TimeSeriesGraph` class.
+
 ## 0.2.9
 
 - Added support for `python` version `3.12`.